requirements.txt
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0